requirements.txt
langchain-openai>=0.3.9
openai
python-dotenv